day-1/clinical-trial-assistant/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.14
chromadb==0.5.4
pypdf==4.2.0
python-dotenv==1.0.1
